factor Client out into its own module
